Add cuewInit() flags to initialize cuda and nvrtc libraries separately
libcuda is part of the driver, and loading it can fail if there is no
CUDA hardware available. libnvrtc can be installed as part of the CUDA
toolkit without a driver, and can be used to compile programs on build
machines that do not have CUDA hardware.

On the other hand, it is also useful to avoid the overhead of loading
libnvrtc when only libcuda is needed.